Updated existing tests for new XML specification
