--- capsul/utils/test/xml_iterative_pipeline.xml
@@ -1,43 +1,0 @@
-<pipeline>
-    <docstring>
-        DARTEL Pipeline
-    </docstring>
-    <processes>
-        <standard name="data_manager">
-            <module>caps.functional_preproc.dartel_template.DataManager</module>
-        </standard>
-        <iterative name="new_segment">
-            <module>caps.utils.segmentation.SPMNewSegment</module>
-            <iter>coregistered_struct_image</iter>
-            <iter>bias_corrected_image</iter>
-            <iter>dartel_input_images</iter>
-            <iter>forward_deformation_field</iter>
-        </iterative>
-        <standard name="dartel">
-            <module>nipype.interfaces.spm.DARTEL</module>
-        </standard>
-    </processes>
-    <links>
-        <link src="new_segment.dartel_input_images" dest="data_manager.images_from_newSegment"/>
-        <link src="data_manager.images" dest="dartel.image_files"/>
-    </links>
-    <inputs>
-        <input name="coregistered_struct_image" dest="new_segment.coregistered_struct_image"/>
-    </inputs>
-    <outputs>
-        <output name="forward_deformation_field" src="new_segment.forward_deformation_field"/>
-        <output name="bias_corrected_image" src="new_segment.bias_corrected_image"/>
-        <output name="output_directory" src="dartel.output_directory"/>
-        <output name="dartel_flow_fields" src="dartel._dartel_flow_fields"/>
-        <output name="final_template_file" src="dartel._final_template_file"/>
-        <output name="template_files" src="dartel._template_files"/>
-    </outputs>
-    <positions>
-        <position process="dartel" x="425" y="-50"/>
-        <position process="data_manager" x="201" y="-25"/>
-        <position process="inputs" x="154" y="-150"/>
-        <position process="outputs" x="737" y="200"/>
-        <position process="new_segment" x="79" y="150"/>
-    </positions>
-    <scale factor="0.5"/> 
-</pipeline>
