--- capsul/utils/test/xml_switch_pipeline.xml
@@ -1,45 +1,0 @@
-<pipeline>
-    <docstring>
-        Auto Generated Pipeline Test
-    </docstring>
-    <processes>
-        <standard name="p1">
-            <module>capsul.utils.test.process.AFunctionToWrap</module>
-        </standard>
-        <standard name="p2">
-            <module>capsul.utils.test.process.AFunctionToWrap</module>
-        </standard>
-        <switch name="switch_p1_p2">
-            <input>p1</input>
-            <input>p2</input>
-            <output>string</output>
-        </switch>
-    </processes>
-    <links>
-        <link src="fname" dest="p2.fname"/>
-        <link src="pdirectory" dest="p2.directory"/>
-        <link src="value" dest="p2.value"/>
-        <link src="enum" dest="p2.enum"/>
-        <link src="list_of_str" dest="p2.list_of_str"/>
-        <link src="p1.string" dest="switch_p1_p2.p1_switch_string"/>
-        <link src="p2.string" dest="switch_p1_p2.p2_switch_string"/>
-    </links>
-    <inputs>
-        <input name="fname" dest="p1.fname"/>
-        <input name="value" dest="p1.value"/>
-        <input name="enum" dest="p1.enum"/>
-        <input name="list_of_str" dest="p1.list_of_str"/>
-        <input name="pdirectory" dest="p1.directory"/>
-    </inputs>
-    <outputs>
-        <output name="out1" src="switch_p1_p2.string"/>
-    </outputs>
-    <positions>
-        <position process="inputs" x="0" y="0"/>
-        <position process="p1" x="200" y="200"/>
-        <position process="p2" x="200" y="-200"/>
-        <position process="switch_p1_p2" x="400" y="0"/>
-        <position process="outputs" x="600" y="0"/>
-    </positions>
-    <scale factor="0.75"/> 
-</pipeline>
